requirements.txt
deepface==0.0.93
opencv-python-headless==4.10.0.84
psycopg2-binary==2.9.10
retina-face==0.0.17
tensorflow==2.10.0
fastapi
uvicorn
python-multipart
jinja2
numpy==1.24.3
